Changes to the Configuration object to properly handle radio actions as well as toggleactions.  Then add Monitoring, Meter falloff, and Meter Hold under the options menu to the Configuration object, so that they will now be saved and recalled between ardour startups as non session specific items.  Also, ardour.rc will be saved now on exit of ardour, wether or not session is saved.  Also some text and color tweaks
git-svn-id: svn://localhost/trunk/ardour2@542 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ gtk2_ardour/editor.h
@@ -296,7 +296,7 @@ class Editor : public PublicEditor
 	void toggle_waveforms_while_recording ();
 	void toggle_measure_visibility ();
 
-	void set_meter_falloff (float);
+	void set_meter_falloff (int);
 	void set_meter_hold (int32_t);
 
 	/* xfades */
